Share the create_litellm_key-recording CLI double via conftest
Problem: ai_keys_test.py's _MintRecordingImbueCloudCli was a near-verbatim
duplicate of agent_creator_test.py's _RecordingImbueCloudCli (identical
create_calls field and create_litellm_key recording body, differing only in
the stub key string).

Fix: move a single RecordingImbueCloudCli next to the shared FakeImbueCloudCli
in desktop_client/conftest.py (the documented home for shared test doubles)
and use it from both test files, dropping the per-file copies and their
now-unused imports.
